Create embeddings directory if it's missing
When style_embeddings is imported, it will load style embeddings from a
directory. It was assuming this directory already exists, but it will
not when running the checked-out repository. This will cause an error
to bubble up to script.js which will break the app very badly since a
lot of critical code will never run.